feat: add builtin error handler and example plugin tests
