+++ src/entity/plugins/prompts/__init__.py
@@ -2,6 +2,5 @@
 
 from .chain_of_thought import ChainOfThoughtPrompt
 from .react import ReActPrompt
-from .basic_error_handler import BasicErrorHandler
 
-__all__ = ["ChainOfThoughtPrompt", "ReActPrompt", "BasicErrorHandler"]
+__all__ = ["ChainOfThoughtPrompt", "ReActPrompt"]
