+++ src/plugins/builtin/__init__.py
@@ -1,1 +1,5 @@
 """Built-in plugins distributed with Entity."""
+
+from .basic_error_handler import BasicErrorHandler
+
+__all__ = ["BasicErrorHandler"]
